+++ ci/bin/verify.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-source $(dirname $0)/terraform.sh
-
-verifyModulesAndPlugins
-formatCheck