+++ examples/runner-public/main.tf
@@ -4,7 +4,7 @@ data "aws_availability_zones" "available" {
 
 module "vpc" {
   source  = "terraform-aws-modules/vpc/aws"
-  version = "2.33"
+  version = "2.70"
 
   name = "vpc-${var.environment}"
   cidr = "10.1.0.0/16"
@@ -21,7 +21,7 @@ module "vpc" {
 
 module "cache" {
   source      = "../../modules/cache"
-  environment = "${var.environment}"
+  environment = var.environment
 }
 
 module "key_pair" {
@@ -122,6 +122,6 @@ resource "null_resource" "cancel_spot_requests" {
 
   provisioner "local-exec" {
     when    = destroy
-    command = "../../ci/bin/cancel-spot-instances.sh ${self.triggers.environment}"
+    command = "../../bin/cancel-spot-instances.sh ${self.triggers.environment}"
   }
 }