+++ kms.tf
@@ -14,3 +14,8 @@ resource "aws_kms_key" "default" {
   )
 }
 
+resource "aws_kms_alias" "default" {
+  count         = var.enable_kms && var.kms_alias_name != "" ? 1 : 0
+  name          = "alias/${var.kms_alias_name}"
+  target_key_id = aws_kms_key.default[0].key_id
+}